Revamp the package to something that can be used
Rather than having an ugly PoC with prints, it now has logging, proper
type hints in itself (excluding upstream bugs) and test-cases showing
off what it does, as well as a demo.

I would not recommend you actually use this in production, as that might
be more fun than you expected it to be, but it actually works, and I'm
sort of impressed that it does.