Implement checking of all loaded modules
This adds a function to also check all already loaded modules for type
hints.  I haven't actually hooked it up to when you load it, and some
exceptions are needed.

Turns out that "mypy" isn't "mypy" clean, neither is tomli that mypy
depends on, so if I enable this on strict, the entire import chain just
collapses and it's hilarious.